[ALPS05245064] p2p: Fix type mismatch
Fix type mismatch

MTK-Commit-Id: b5fa27ac582e1574cc09b4dd4a801fa05c00a08b

CR-Id: ALPS05245064
Feature: [Module] Wi-Fi Driver

+++ os/linux/gl_p2p.c
@@ -1108,7 +1108,7 @@ u_int8_t p2pNetUnregister(struct GLUE_INFO *prGlueInfo,
  */
 /*---------------------------------------------------------------------------*/
 int glSetupP2P(struct GLUE_INFO *prGlueInfo, struct wireless_dev *prP2pWdev,
-		struct net_device *prP2pDev, int u4Idx, u_int8_t fgIsApMode)
+		struct net_device *prP2pDev, uint8_t u4Idx, u_int8_t fgIsApMode)
 {
 	struct ADAPTER *prAdapter = NULL;
 	struct GL_P2P_INFO *prP2PInfo = NULL;
